// public/angular/app/interfaces/INotifDevuelta.ts
export interface INotifDevuelta {
    devuelta: number;
    notificacion: number;
    empresa_mensajeria: number;
    numero_guia: string;
    fecha_novedad: string;
    causal_devolucion: number;
    observaciones: string;
    fecha_registra: string;
    usuario: number;
    usuario_desc: string;
}