// public/angular/app/interfaces/IOpcion.ts
export interface IOpcion {
    modulo: number;
    modulo_nombre: string;
    opcion: number;
    opcion_nombre: string;
    enlace: string;
    tipo: string;
    permiso: number;
    consulta: number;
    adiciona: number;
    edita: number;
    elimina: number;
    ejecuta: number;
    fecha_desde: number;
    fecha_hasta: number;
}